Implement load favicon task

+++ bookmarks/services/favicon_loader.py
@@ -8,13 +8,13 @@ import requests
 from django.conf import settings
 
 
+def _ensure_favicon_folder():
+    Path(settings.LD_FAVICON_FOLDER).mkdir(parents=True, exist_ok=True)
+
+
 def _url_to_filename(url: str) -> str:
     name = re.sub(r'\W+', '_', url)
     return f'{name}.png'
-
-
-def _ensure_favicon_folder():
-    Path(settings.LD_FAVICON_FOLDER).mkdir(parents=True, exist_ok=True)
 
 
 def _get_base_url(url: str) -> str:
@@ -22,28 +22,27 @@ def _get_base_url(url: str) -> str:
     return f'{parsed_uri.scheme}://{parsed_uri.hostname}'
 
 
-def _get_favicon_path(base_url: str) -> str:
-    favicon_name = _url_to_filename(base_url)
-    return os.path.join(settings.LD_FAVICON_FOLDER, favicon_name)
+def _get_favicon_path(favicon_file: str) -> str:
+    return os.path.join(settings.LD_FAVICON_FOLDER, favicon_file)
 
 
-def check_favicon(url: str) -> bool:
-    base_url = _get_base_url(url)
-    favicon_path = _get_favicon_path(base_url)
-    return Path(favicon_path).exists()
-
-
-def load_favicon(url: str):
-    # Create favicon folder if not exists
-    _ensure_favicon_folder()
+def load_favicon(url: str, force_update=False) -> str:
     # Get base URL so that we can reuse favicons for multiple bookmarks with the same host
     base_url = _get_base_url(url)
-    # Load favicon from provider, save to file
-    favicon_url = settings.LD_FAVICON_PROVIDER.format(url=base_url)
-    favicon_path = _get_favicon_path(base_url)
-    response = requests.get(favicon_url, stream=True)
+    favicon_name = _url_to_filename(base_url)
+    favicon_path = _get_favicon_path(favicon_name)
 
-    with open(favicon_path, 'wb') as file:
-        shutil.copyfileobj(response.raw, file)
+    # Load icon if it doesn't exist yet
+    if not Path(favicon_path).exists() or force_update:
+        # Create favicon folder if not exists
+        _ensure_favicon_folder()
+        # Load favicon from provider, save to file
+        favicon_url = settings.LD_FAVICON_PROVIDER.format(url=base_url)
+        response = requests.get(favicon_url, stream=True)
 
-    del response
+        with open(favicon_path, 'wb') as file:
+            shutil.copyfileobj(response.raw, file)
+
+        del response
+
+    return favicon_name
